+++ tests/sim/simulation/actions/material/insert_material/test_insert_random_material.py
@@ -11,7 +11,7 @@ class TestInsertRandomMaterial(TestCase):
         # arrange
         simulation = WarehouseSimulation(Warehouse())
         duration = 5
-        init_num_materials: int = simulation.get_warehouse().get_carousel().get_deposit_tray().get_num_materials()
+        init_num_materials: int = simulation.get_warehouse().get_carousel().get_bay_tray().get_num_materials()
         material = InsertRandomMaterial(simulation.get_environment(), simulation.get_warehouse(), simulation, duration)
 
         # act
@@ -24,4 +24,4 @@ class TestInsertRandomMaterial(TestCase):
         self.assertIsInstance(action_performed, dict)
         self.assertEqual(ActionEnum.INSERT_RANDOM_MATERIAL.value, action_performed['Type of Action'])
         self.assertGreaterEqual((action_performed['Finish']-action_performed['Start']).total_seconds(), duration)
-        self.assertGreater(simulation.get_warehouse().get_carousel().get_deposit_tray().get_num_materials(), init_num_materials)
+        self.assertGreater(simulation.get_warehouse().get_carousel().get_bay_tray().get_num_materials(), init_num_materials)